WIP refactor header
