+++ src/components/header/navbar/MobileNavbar.jsx
@@ -1,0 +1,46 @@
+import React, { useState, useEffect } from 'react'
+import {
+  CardMedia,
+  Box,
+  Grid,
+  useMediaQuery,
+  SwipeableDrawer,
+  IconButton,
+} from '@mui/material'
+import MenuIcon from '@mui/icons-material/Menu'
+import useStyles from '../../../theme/views/navbar.theme'
+import Adaptive_logo from '../../../assets/images/Adaptive_logo_WHITE_PINK.png'
+
+const MobileNavbar = ({tabs}) => {
+  const classes = useStyles()
+  const [drawerOpen, setDrawerOpen] = useState(false)
+
+  return (
+    <Box data-cy='navbar' className={classes.container}>
+      <img
+        className={classes.logo}
+        src={Adaptive_logo}
+        data-cy='adaptive-logo'
+        alt='The Adaptive Logo'
+      />
+
+      <SwipeableDrawer
+        data-cy='navbar'
+        PaperProps={{ className: classes.drawer }}
+        anchor='right'
+        open={drawerOpen}
+        onClose={() => setDrawerOpen(false)}
+        onOpen={() => setDrawerOpen(true)}
+        onClick={() => setDrawerOpen(false)}>
+        {tabs}
+      </SwipeableDrawer>
+      <IconButton
+        className={classes.menuButton}
+        onClick={() => setDrawerOpen(true)}>
+        <MenuIcon data-cy='hamburger-menu' />
+      </IconButton>
+    </Box>
+  )
+}
+
+export default MobileNavbar
